Fix recette dashboardv2 (#2893)
* fix: operator_graph

* fix: api dashboard

+++ api/src/pdc/services/dashboard/providers/CampaignsRepository.ts
@@ -56,8 +56,8 @@ export class CampaignsRepository implements CampaignsRepositoryInterface {
         a.unit,
         a.status,
         a.handler,
-        a.incentive_sum::int,
-        a.max_amount::int 
+        a.incentive_sum,
+        a.max_amount 
       FROM ${raw(this.table)} a
       LEFT JOIN ${raw(this.tableTerritory)} b on a.territory_id = b._id
       ${params.operator_id ? sql`LEFT JOIN ${raw(this.tableIncentives)} c on a._id = c.policy_id` : sql``}
